+++ src/main/java/org/springframework/core/io/ClassPathResource.java
@@ -1,0 +1,28 @@
+package org.springframework.core.io;
+
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Objects;
+
+/**
+ * @author derekyi
+ */
+public class ClassPathResource implements Resource {
+
+    private final String path;
+
+    public ClassPathResource(String path) {
+        this.path = path;
+    }
+
+    @Override
+    public InputStream getInputStream() throws IOException {
+        InputStream is = this.getClass().getClassLoader().getResourceAsStream(path);
+        if (Objects.isNull(is)) {
+            throw new FileNotFoundException(path + " cannot be opened because it does not exist");
+        }
+        return is;
+    }
+
+}