+++ src/main/java/org/springframework/core/io/FileSystemResource.java
@@ -1,0 +1,25 @@
+package org.springframework.core.io;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+/**
+ * @author derekyi
+ */
+public class FileSystemResource implements Resource {
+
+    private final String filePath;
+
+    public FileSystemResource(String filePath) {
+        this.filePath = filePath;
+    }
+
+    @Override
+    public InputStream getInputStream() throws IOException {
+        Path path = new File(filePath).toPath();
+        return Files.newInputStream(path);
+    }
+}